requirements.txt
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.16.0 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
greenlet==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
grpcio-tools==1.62.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.62.1 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
outcome==1.3.0.post0 ; python_version >= "3.12" and python_version < "4.0"
overrides==7.7.0 ; python_version >= "3.12" and python_version < "4.0"
playwright==1.43.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.12" and python_version < "4.0"
pyee==11.1.0 ; python_version >= "3.12" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.12" and python_version < "4.0"
robotframework-assertion-engine==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
robotframework-browser==18.3.0 ; python_version >= "3.12" and python_version < "4.0"
robotframework-pythonlibcore==4.4.1 ; python_version >= "3.12" and python_version < "4"
robotframework==7.0 ; python_version >= "3.12" and python_version < "4.0"
selenium==4.19.0 ; python_version >= "3.12" and python_version < "4.0"
setuptools==69.5.1 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
trio-websocket==0.11.1 ; python_version >= "3.12" and python_version < "4.0"
trio==0.25.0 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3[socks]==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.12" and python_version < "4.0"
